Extract Lightbox component from ImgController
